streamlit/requirements.txt
streamlit
requests
pandas
bcrypt
python-dotenv